Fix to return size_t from strlen

+++ commonsrc/commonsrc.h
@@ -8,7 +8,7 @@ char* strncpy(char*, const char*, size_t);
 char* strchr(const char*, char);
 char* strrchr(const char*, char);
 int strncmp(const char*, const char*, size_t);
-int strlen(const char*);
+size_t strlen(const char*);
 
 int memcmp(const void *v1, const void *v2, size_t n);
 void* memcpy(void *dst, const void *src, size_t n);
